*added post share route and error handling for comments *created update profile and cover image routes

+++ src/controllers/postController.js
@@ -1,23 +1,26 @@
 import AppError from '../lib/AppError.js';
 import { asyncWrapper } from '../lib/asyncWrapper.js';
+
 import User from '../models/User.js';
 import Post from '../models/Post.js';
 import Comment from '../models/Comment.js';
+
 import fs from 'fs';
 import { promisify } from 'util';
 
 import { uploadMedia, editMedia } from '../lib/multer.js';
 
-export const uploadImages = uploadMedia({
-  storage: 'memoryStorage',
-  upload: 'any',
-  filename: 'images',
-});
-
 export const resizeAndOptimiseMedia = editMedia({
   multy: true,
   resize: false,
 });
+
+export const uploadPostMediaFiles = (imageName) =>
+  uploadMedia({
+    storage: 'memoryStorage',
+    upload: 'any',
+    filename: imageName,
+  });
 
 export const createPost = asyncWrapper(async function (req, res, next) {
   const { type, description } = req.body;
@@ -45,7 +48,7 @@ export const createPost = asyncWrapper(async function (req, res, next) {
 
   await newPost.save();
 
-  res.status(200).json(newPost);
+  res.status(201).json(newPost);
 });
 
 export const deletePost = asyncWrapper(async function (req, res, next) {
@@ -59,7 +62,7 @@ export const deletePost = asyncWrapper(async function (req, res, next) {
 
   const postMedia = postToDelete.media;
 
-  if (postMedia && postMedia.length > 0) {
+  if (!postToDelete.shared && postMedia && postMedia.length > 0) {
     const deletion = promisify(fs.unlink);
 
     Promise.all(
@@ -70,14 +73,16 @@ export const deletePost = asyncWrapper(async function (req, res, next) {
         } catch (error) {
           return next(
             new AppError(
-              403,
-              "something went wrong, cant't delete post media files please try again"
+              406,
+              "something went wrong, cant't find and delete post media files which are attached to your post. please report the problem or try later"
             )
           );
         }
       })
     );
   }
+
+  await Comment.deleteMany({ post: postToDelete._id });
 
   await postToDelete.delete();
 
@@ -89,19 +94,10 @@ export const updatePost = asyncWrapper(async function (req, res, next) {
   const { description, media } = req.body;
   const currUser = req.user;
 
-  const post = await Post.findById(postId)
-    .populate({
-      path: 'author',
-      select: 'userName profileImg',
-    })
-    .populate({
-      path: 'authenticAuthor',
-      select: 'userName profileImg',
-    })
-    .populate({
-      path: 'reactions.author',
-      select: 'userName profileImg',
-    });
+  const post = await Post.findById(postId).populate({
+    path: 'author authenticAuthor reactions.author',
+    select: 'userName profileImg',
+  });
 
   if (!post || post.author._id.toString() !== currUser.id)
     return next(new AppError(404, 'post does not exists'));
@@ -110,7 +106,7 @@ export const updatePost = asyncWrapper(async function (req, res, next) {
 
   const existingFiles = post.media;
   const filteredMedia = [];
-  if (existingFiles?.[0])
+  if (!post.shared && existingFiles?.[0])
     Promise.all(
       existingFiles.map(async (file) => {
         try {
@@ -122,14 +118,14 @@ export const updatePost = asyncWrapper(async function (req, res, next) {
           return next(
             new AppError(
               403,
-              "something went wrong, cant't delete removed post media files please try again"
+              "something went wrong, cant't find and delete removed post media files which are attached to your post.  please report the problem or try later"
             )
           );
         }
       })
     );
 
-  if (req.files) {
+  if (!post.shared && req.files) {
     const newFiles = req.xOriginal.map(
       (fileName) => `${req.protocol}://${'localhost:4000'}/${fileName}`
     );
@@ -140,13 +136,13 @@ export const updatePost = asyncWrapper(async function (req, res, next) {
     // const match = await matchModifiedFilesToExisting(`public/images/${originalFileName}`);
 
     post.media = [...modifiedExistingFiles, ...newFiles];
-  } else post.media = media;
+  } else if (!post.shared) post.media = media;
 
   if (description) post.description = description;
 
   await post.save();
 
-  res.status(200).json(post);
+  res.status(201).json(post);
 });
 
 export const reactOnPost = asyncWrapper(async function (req, res, next) {
@@ -155,6 +151,8 @@ export const reactOnPost = asyncWrapper(async function (req, res, next) {
   const currUser = req.user;
 
   const post = await Post.findById(postId);
+
+  if (!post) return next(new AppError(404, 'post does not exists'));
 
   const existingReaction = post.reactions.find(
     (reaction) => reaction.author.toString() === currUser.id
@@ -184,29 +182,13 @@ export const reactOnPost = asyncWrapper(async function (req, res, next) {
 export const getPostComments = asyncWrapper(async function (req, res, next) {
   const { postId } = req.params;
 
+  const post = await Post.findById(postId);
+
+  if (!post) return next(new AppError(404, 'post does not exists'));
+
   const comments = await Comment.find({ post: postId })
     .populate({
-      path: 'author',
-      select: 'userName profileImg',
-    })
-    .populate({
-      path: 'tags',
-      select: 'userName profileImg',
-    })
-    .populate({
-      path: 'reactions.author',
-      select: 'userName profileImg',
-    })
-    .populate({
-      path: 'replies.author',
-      select: 'userName profileImg',
-    })
-    .populate({
-      path: 'replies.reactions.author',
-      select: 'userName profileImg',
-    })
-    .populate({
-      path: 'replies.tags',
+      path: 'author tags reactions.author replies.author replies.reactions.author replies.tags',
       select: 'userName profileImg',
     })
     .sort({ createdAt: -1 });
@@ -214,14 +196,52 @@ export const getPostComments = asyncWrapper(async function (req, res, next) {
   res.status(200).json(comments);
 });
 
-/////////////////////////////////////////////////////////////////////
-export const getPost = asyncWrapper(async function (req, res, next) {
-  const post = await Post.findById(postId).populate('author').populate('authenticAuthor');
+export const sharePost = asyncWrapper(async function (req, res, next) {
+  const { postId } = req.params;
+  const { description } = req.body;
+  const currUser = req.user;
 
-  if (!post) throw new Error('there are no such a post');
+  const postToShare = await Post.findById(postId);
 
-  res.status(200).json();
+  const sharedPost = {
+    type: 'post',
+    author: currUser.id,
+    description: description,
+    media: postToShare.media,
+    shared: true,
+    authenticType: postToShare.shared ? postToShare.authenticType : postToShare.type,
+    authenticAuthor: postToShare.shared ? postToShare.authenticAuthor : postToShare.author,
+    authenticDescription: postToShare.shared
+      ? postToShare.authenticDescription
+      : postToShare.description,
+    authenticDateCreation: postToShare.shared
+      ? postToShare.authenticDateCreation
+      : postToShare.createdAt,
+  };
+
+  const newPost = await Post.create(sharedPost);
+
+  await newPost.populate({
+    path: 'author authenticAuthor',
+    select: 'userName profileImg',
+  });
+
+  res.status(201).json(newPost);
 });
+
+export const getPost = asyncWrapper(async function (req, res, next) {
+  const { postId } = req.params;
+
+  const post = await Post.findById(postId)
+    .populate({ path: 'author', select: 'userName profileImg' })
+    .populate({ path: 'authenticAuthor', select: 'userName profileImg' });
+
+  if (!post) return next(new AppError(404, 'post does not exists'));
+
+  res.status(200).json(post);
+});
+
+/////////////////////////////////////////////////////////////////////
 
 export const getAllPosts = asyncWrapper(async function (req, res, next) {
   const posts = await Post.find()
@@ -234,26 +254,6 @@ export const getAllPosts = asyncWrapper(async function (req, res, next) {
       path: 'comments',
       populate: { path: 'author replies.author replies.adressat' },
     });
-
-  res.status(200).json();
-});
-
-export const sharePost = asyncWrapper(async function (req, res, next) {
-  const postToShare = await Post.findById(postId);
-
-  const sharedPost = {
-    type: 'post',
-    author: currUser.id,
-    description: description,
-    media: postToShare.media,
-    shared: true,
-    authenticType: postToShare.type,
-    authenticAuthor: postToShare.author,
-    authenticDescription: postToShare.description,
-    authenticDateCreation: postToShare.createdAt,
-  };
-
-  const newPost = await Post.create(sharedPost);
 
   res.status(200).json();
 });
